Made select statements log properly for everyone but SystemUser.
SystemUser was breaking the activity log by creating a circular dependency.
Since we load the current user in order to log who did what, that triggers the AfterLoad
which in turn breaks our system.  We may need to create an individual Subscriber for just
the SystemUser and have all the other Entities use a shared one.