AWS Authentication Backend Verification, TODO fixes.
Made it so the AWS authentication was actually being validated on the backend.  The issued timestamp is verified as well as making sure that only the mobile app audience is allowed in.  The JWT encryption signature is verified against the AWS payload.

Also found a bug where we were creating multiple users over and over again when logging in.  That's now fixed as the identityID on the backend checks against the database identity.

Minor cleanup on the frontend.